requirements.txt
requests
pandas~=1.4.3
tabulate~=0.8.10
anvil-uplink
jsbeautifier
numpy~=1.23.2
pymongo